82439-code changes for UoS validation and duplicate price information AB#82439 AB#82672
