--- src/UKHO.ERPFacade.API/Services/IERPFacadeService.cs
@@ -1,12 +1,0 @@
-﻿using Newtonsoft.Json.Linq;
-using UKHO.ERPFacade.API.Models;
-
-namespace UKHO.ERPFacade.API.Services
-{
-    public interface IErpFacadeService
-    {
-        List<UnitsOfSalePrices> MapAndBuildUnitsOfSalePrices(List<PriceInformation> priceInformationList, List<UnitOfSale> encEventUnitOfSaleList);
-
-        UnitOfSaleUpdatedEventPayload BuildUnitsOfSaleUpdatedEventPayload(List<UnitsOfSalePrices> unitsOfSalePriceList, string encEventPayloadJson);        
-    }
-}
