+++ src/UKHO.ERPFacade.Common/Models/PriceInformationEvent.cs
@@ -1,10 +1,10 @@
 ﻿using Newtonsoft.Json;
 using System.Diagnostics.CodeAnalysis;
 
-namespace UKHO.ERPFacade.API.Models
+namespace UKHO.ERPFacade.Common.Models
 {
     [ExcludeFromCodeCoverage]
-    public class PriceInformation 
+    public class PriceInformation
     {
         [JsonProperty("corrid")]
         public string Corrid { get; set; }
