+++ src/UKHO.ERPFacade.Common/Models/UnitOfSaleUpdatedEventPayload.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using UKHO.ERPFacade.Common.Infrastructure;
 
-namespace UKHO.ERPFacade.API.Models
+namespace UKHO.ERPFacade.Common.Models
 {
     [ExcludeFromCodeCoverage]
     public class UnitOfSaleUpdatedEventPayload : EventBase<UnitOfSaleUpdatedEvent>
