+++ src/UKHO.ERPFacade.Common/Services/IERPFacadeService.cs
@@ -1,0 +1,11 @@
+﻿using UKHO.ERPFacade.Common.Models;
+
+namespace UKHO.ERPFacade.Common.Services
+{
+    public interface IErpFacadeService
+    {
+        List<UnitsOfSalePrices> MapAndBuildUnitsOfSalePrices(List<PriceInformation> priceInformationList, List<string> unitOfSalesList);
+
+        UnitOfSaleUpdatedEventPayload BuildUnitsOfSaleUpdatedEventPayload(List<UnitsOfSalePrices> unitsOfSalePriceList, string encEventPayloadJson);
+    }
+}
